feat(wip): new architecure for v4
