+++ packages/api/src/typeDefs/advertisement.graphql
