+++ packages/api/src/typeDefs/aggregateOffer.graphql
