+++ packages/api/src/typeDefs/aggregateRating.graphql
