+++ packages/api/src/typeDefs/image.graphql
