+++ packages/api/src/typeDefs/newsletter.graphql
