+++ packages/api/src/typeDefs/product.graphql
