+++ packages/api/src/typeDefs/productGroup.graphql
