+++ packages/core/@generated/index.ts
@@ -1,1 +1,1 @@
-export * from "./gql";
+export * from './gql'
