--- packages/core/output/handlers.ts
@@ -1,4 +1,0 @@
-export { default as GraphqlHandler } from '../src/pages/api/graphql'
-export { default as PreviewHandler } from '../src/pages/api/preview'
-export { default as LiveHandler } from '../src/pages/api/health/live'
-export { default as ReadyHandler } from '../src/pages/api/health/ready'
