+++ packages/core/src/utils/withFaststore.js
@@ -158,3 +158,5 @@ function mergeConfig() {
 
   return deepmerge(baseDiscoveryConfig, faststoreConfig)
 }
+
+/** @typedef {(next: import('next').NextConfig['webpack']) => import('next').NextConfig['webpack']} AnotateWebpack */
