--- packages/graphql-schema/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "@faststore/graphql-schema",
-  "version": "1.0.0",
-  "description": "",
-  "bin": {
-    "faststore-schema": "./dist/bin.js"
-  },
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "default": "./dist/index.js"
-    },
-    "./schema.graphql": "./dist/schema.graphql"
-  },
-  "scripts": {
-    "build": "vite build",
-    "schema:gen": "pnpm run build && pnpm dlx tsx ./generate-schema.ts ./dist/schema.graphql",
-    "dev": "pnpm run schema:gen --watch"
-  },
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@types/glob-watcher": "^5.0.5",
-    "@types/node": "^24.2.1",
-    "glob-watcher": "^6.0.0",
-    "typescript": "5.3.2",
-    "vite": "^7.1.2",
-    "vite-plugin-dts": "^4.5.4",
-    "@graphql-tools/merge": "^9.1.1",
-    "graphql": "^15.8.0",
-    "graphql-tag": "^2.12.6"
-  },
-  "peerDependencies": {},
-  "dependencies": {
-    "@faststore/graphql-schema": "workspace:^"
-  }
-}
